+++ adapter/src/main/java/ee/cyber/cdoc2/server/adapter/rest/MidRestClient.java
@@ -3,7 +3,9 @@ package ee.cyber.cdoc2.server.adapter.rest;
 import ee.sk.mid.MidClient;
 import ee.sk.mid.MidDisplayTextFormat;
 import ee.sk.mid.MidLanguage;
+import ee.sk.mid.rest.dao.MidSessionStatus;
 import ee.sk.mid.rest.dao.request.MidAuthenticationRequest;
+import ee.sk.mid.rest.dao.request.MidSessionStatusRequest;
 import ee.sk.mid.rest.dao.response.MidAuthenticationResponse;
 import lombok.RequiredArgsConstructor;
 
@@ -13,15 +15,16 @@ import org.springframework.stereotype.Component;
 
 import ee.cyber.cdoc2.server.adapter.conf.MobileIdClientConf;
 import ee.cyber.cdoc2.server.app.usecase.startauth.MidAuthenticate;
+import ee.cyber.cdoc2.server.app.usecase.status.mid.GetMidSession;
 
 @Component
 @RequiredArgsConstructor
-public class MidRestClient implements MidAuthenticate {
+public class MidRestClient implements MidAuthenticate, GetMidSession {
     private final MidClient midClient;
     private final MobileIdClientConf.AppProperties properties;
 
     @Override
-    public UUID execute(Request request) {
+    public UUID execute(MidAuthenticate.Request request) {
         MidAuthenticationRequest authenticationRequest = MidAuthenticationRequest.newBuilder()
             .withPhoneNumber(request.phoneNumber())
             .withNationalIdentityNumber(request.nationalIdentityNumber())
@@ -37,5 +40,23 @@ public class MidRestClient implements MidAuthenticate {
             .authenticate(authenticationRequest);
 
         return UUID.fromString(response.getSessionID());
+    }
+
+    @Override
+    public Response execute(UUID sessionId) {
+        MidSessionStatus midSessionStatus = midClient.getMobileIdConnector()
+            .getAuthenticationSessionStatus(
+                new MidSessionStatusRequest(
+                    sessionId.toString(),
+                    properties.timeoutSeconds()
+                )
+            );
+
+        return new Response(
+            midSessionStatus.getState(),
+            midSessionStatus.getResult(),
+            midSessionStatus.getCert(),
+            midSessionStatus.getSignature()
+        );
     }
 }